communication details code added
