--- erp.lt.portal/src/main/java/com/erp/lt/portal/constant/ERPConstant.java
@@ -1,5 +1,0 @@
-package com.erp.lt.portal.constant;
-
-public class ERPConstant {
-
-}
