--- erp.lt.portal/src/main/java/com/erp/lt/portal/repository/EmployeeEduRepo.java
@@ -1,10 +1,0 @@
-package com.erp.lt.portal.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.erp.lt.portal.dto.EducationDetailsDto;
-@Repository
-public interface EmployeeEduRepo extends JpaRepository<EducationDetailsDto, Long> {
-
-}
