+++ erp.lt.portal/src/main/java/com/erp/lt/portal/service/MobileDetailsService.java
@@ -1,0 +1,18 @@
+/**
+ * 
+ */
+package com.erp.lt.portal.service;
+
+import com.erp.lt.portal.vo.MobileDetailsVO;
+
+/**
+ * @author User
+ *
+ */
+public interface MobileDetailsService {
+	public MobileDetailsVO getMobileDetails( int code);
+	public void addMobileDetials(MobileDetailsVO mobileDetailsVO);
+	public void editMobileDetials(MobileDetailsVO mobileDetailsVO);
+	
+	public void deleteMobileDetailsByComCode(int code);
+}
