--- erp.lt.portal/src/main/java/com/erp/lt/portal/service/PreviousEmployerService.java
@@ -1,24 +1,0 @@
-/**
- * 
- */
-package com.erp.lt.portal.service;
-
-import java.util.Optional;
-
-import com.erp.lt.portal.dto.PreviousEmployeeVO;
-
-/**
- * @author User
- *
- */
-public interface PreviousEmployerService {
-	
-	public void addPreviousEmpDetail(PreviousEmployeeVO previousEmployeeVO);
-
-	public void editPreviousEmpDetail(PreviousEmployeeVO previousEmployeeVO);
-
-	public void deletePreviousEmpDetail(Long empid);
-
-	public Optional<PreviousEmployeeVO>getPreviousEmpDetail(Long empid);
-
-}
