+++ erp.lt.portal/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 server.port=8081
-spring.application.name=/ErpPorts
+spring.application.name=//erp.lt.portal
 #spring.datasource.url=jdbc:h2:mem:testdb
 #spring.datasource.driverClassName=org.h2.Driver
 #spring.datasource.username=sa
@@ -21,7 +21,7 @@ spring.insights.web.log-request-details=true
 
 debug = true
 spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mysql://localhost:53306/laderaslides?useSSL=true&
+spring.datasource.url=jdbc:mysql://localhost:53306/employee_selfservicedb?useSSL=true&
 spring.datasource.username=root
 spring.datasource.password=root
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
