--- erp.lt.portal/src/test/java/com/erp/lt/portal/ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.erp.lt.portal;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
